Lesson 18. Exception handling will be added to the entire project

+++ spring/spring-java20-jahangir-askerov/src/main/java/az/developia/spring_java20_jahangir_askerov/service/BookService.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
 
+import az.developia.spring_java20_jahangir_askerov.exception.NotFoundException;
 import az.developia.spring_java20_jahangir_askerov.model.BookEntity;
 import az.developia.spring_java20_jahangir_askerov.model.BookUpdateRequest;
 import az.developia.spring_java20_jahangir_askerov.repository.BookRepository;
@@ -28,24 +29,30 @@ public class BookService {
 	public BookEntity getBookById(Integer id) {
 		Optional<BookEntity> optionalBook = repository.findById(id);
 		if (optionalBook.isEmpty()) {
-			return null;
+			throw new NotFoundException("The book with the ID you are looking for does not exist !");
 		}
 
 		return optionalBook.get();
 	}
 
 	public void deleteBookById(Integer id) {
+		Optional<BookEntity> optionalBook = repository.findById(id);
+		if(optionalBook.isEmpty()) {
+			throw new NotFoundException("The book with the ID you are looking for does not exist !");
+		}
+		
 		repository.deleteById(id);
 	}
 
 	public void updateBookById(Integer id, BookUpdateRequest book) {
 		Optional<BookEntity> optionalBook = repository.findById(id);
-		BigDecimal price = book.getPrice();
-		String description = book.getDescription();
 		if (optionalBook.isEmpty()) {
-			return;
+			throw new NotFoundException("The book with the ID you are looking for does not exist !");
 		}
 
+		BigDecimal price = book.getPrice();
+		String description = book.getDescription();
+		
 		BookEntity existingBook = optionalBook.get();
 		existingBook.setPrice(price);
 		existingBook.setDescription(description);
